:lipstick: remove unused imports
