+++ app/src/api/arrangementSvc.ts
@@ -1,11 +1,7 @@
 import { IEvent, serializeEvent, IEventContract } from 'src/types/event';
 import { post, get, del, put } from './crud';
 import { WithId } from 'src/types';
-import {
-  IEditParticipant,
-  IParticipantContract,
-  IParticipant,
-} from 'src/types/participant';
+import { IParticipantContract, IParticipant } from 'src/types/participant';
 
 const host = 'https://api.dev.bekk.no/arrangement-svc/';
