--- app/src/components/Common/Menu/Menu.module.scss
@@ -1,17 +1,0 @@
-.container {
-  display: flex;
-  justify-content: space-between;
-}
-
-.menuItem {
-  text-align: left;
-  text-decoration: none;
-  font-size: 24px;
-  &:hover {
-    cursor: pointer;
-  }
-}
-
-.selected {
-  text-decoration: underline;
-}
