+++ app/src/components/Common/TimeInput/TimeInput.tsx
@@ -3,12 +3,11 @@ import { EditTime } from 'src/types/time';
 import style from './TimeInput.module.scss';
 
 interface IProps {
-  label: string;
   value: EditTime;
   onChange: (value: EditTime) => void;
 }
 
-export const TimeInput = ({ label, value, onChange }: IProps): JSX.Element => {
+export const TimeInput = ({ value, onChange }: IProps): JSX.Element => {
   const hour = value[0];
   const minute = value[1];
 
@@ -25,14 +24,12 @@ export const TimeInput = ({ label, value, onChange }: IProps): JSX.Element => {
       <input
         className={style.timeInput}
         type="number"
-        id={label}
         onChange={v => updateHour(v.target.value)}
         value={hour}
       />
       <input
         className={style.timeInput}
         type="number"
-        id={label}
         onChange={v => updateMinute(v.target.value)}
         value={minute}
       />
