--- app/src/components/CreateEvent/CreateEventContainer.module.scss
@@ -1,20 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.subsection {
-  margin: 20px 0px;
-}
-
-.form {
-  display: flex;
-  flex-direction: column;
-}
-
-.dateTime {
-  display: flex;
-  flex-direction: row;
-}
